+++ packages/ui/CHANGELOG.md
@@ -9,6 +9,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Added
 
 - `ui/Button`, `ui/Icon`, and `ui/IconButton` components to support reuse by themes
+- `ui/Touchable` support for flick gestures
 
 ### Changed